Lazy Configuration with fallback for Ant Task
Currently, the Ant Task is evaluated during the Configuration phase,
and it is not possible to add a dynamic configuration.
With this change the ant properties are lazily evaluated, which allows to eg use outcomes of other tasks.
Additionally, we provide a fallback - if a property is not set.
This allows to define placeholders within the project.properties and a fallback if needed. Useful when you want to add the jacoco agent from gradle to the standalone.javaoptions.